Update Swagger file (#44)
